requirements.txt
Django==5.1.7
pandas==2.2.3
seaborn==0.13.2
matplotlib==3.10.0
numpy==2.2.3
django-pandas==0.6.7
wordcloud==1.9.4
asgiref                       3.8.1
bcrypt                        4.3.0
CacheControl                  0.14.3
cachetools                    5.5.2
certifi                       2025.4.26
cffi                          1.17.1
charset-normalizer            3.4.2
contourpy                     1.3.2
cryptography                  44.0.2
cycler                        0.12.1
defusedxml                    0.7.1
Django                        5.0.14
django-appconf                1.1.0
django-cors-headers           4.7.0
django-extensions             3.2.3
django-mssql-backend          2.8.1
django-pandas                 0.6.7
django-select2                8.4.0
djangorestframework           3.14.0
djangorestframework_simplejwt 5.5.0
et_xmlfile                    2.0.0
firebase-admin                6.8.0
fonttools                     4.57.0
google-api-core               2.25.0rc0
google-api-python-client      2.169.0
google-auth                   2.39.0
google-auth-httplib2          0.2.0
google-auth-oauthlib          1.2.2
google-cloud-core             2.4.3
google-cloud-firestore        2.20.2
google-cloud-storage          3.1.0
google-crc32c                 1.7.1
google-resumable-media        2.7.2
googleapis-common-protos      1.70.0
grpcio                        1.71.0
grpcio-status                 1.71.0
httplib2                      0.22.0
idna                          3.10
kiwisolver                    1.4.8
matplotlib                    3.10.1
msgpack                       1.1.0
mssql-django                  1.5
numpy                         2.2.5
oauthlib                      3.2.2
openpyxl                      3.1.5
packaging                     25.0
pandas                        2.2.3
pillow                        11.2.1
pip                           25.0.1
proto-plus                    1.26.1
protobuf                      5.29.4
psycopg2                      2.9.10
psycopg2-binary               2.9.10
pyasn1                        0.6.1
pyasn1_modules                0.4.2
pycparser                     2.22
PyJWT                         2.9.0
pymongo                       3.11.4
pyodbc                        5.2.0
pyparsing                     3.2.3
python-dateutil               2.9.0.post0
python-dotenv                 1.1.0
python3-openid                3.2.0
pytz                          2025.1
requests                      2.32.3
requests-oauthlib             2.0.0
rsa                           4.9.1
seaborn                       0.13.2
setuptools                    78.1.0
six                           1.17.0
social-auth-app-django        5.4.3
social-auth-core              4.6.1
sqlparse                      0.5.3
tzdata                        2025.1
uritemplate                   4.1.1
urllib3                       2.4.0
wordcloud                     1.9.4